Remove Statuses Limitations In The Dashboard
The dashboard was unable to display statuses other than build, test
and deliver. This happened due to go template limitations on map
structures. To surpass those limitations, a Statuses array was added in
the DeployElement struct, allowing Highcharts to better display the
necessary data and enabling developers to use any desired status.